created prefab of entire demo scene
still need to decide on connection between grid & pillar. ask Tre what he feels the best way would be based on plan for future code.
demo scene prefab is for easy transfer between scenes. May not be necessary and can be deleted if not used to save space -- also if needed.